fix CI
bug introduced with #5370

ci: full-compile

+++ include/picongpu/particles/densityProfiles/GaussianImpl.hpp
@@ -47,7 +47,7 @@ namespace picongpu
              */
             HDINLINE float_X operator()(DataSpace<simDim> const& totalCellOffset)
             {
-                if(totalCellOffset.y() < ParamClass::vacuumCellsY)
+                if(static_cast<uint32_t>(totalCellOffset.y()) < ParamClass::vacuumCellsY)
                 {
                     return 0._X;
                 }
